Resolved the ambiguities in the parser.
'=' is now treated the same as '|' in productions. This means that this:

<Foo> = <Bar>
<Baz> = <Quux>

is turned into an AST like this (<Foo> = <Bar> <Baz>) (= <Quux>). We'll need a post-processing stage to remove the final nonterminals from items.

This also won't see things like foo = <Bar> as an error except for in the first item. This is perhaps not so great.

--HG--
branch : Parse
extra : convert_revision : svn%3A3666fe7e-a552-0410-bc3c-bd4e4c681ee0/trunk/Parse%401064