Added two new actions: divert and guard.
Divert moves to a new state without affecting the lookahead. It's useful for cases where a guard symbol is not in conflict with another action.

Guard pauses the current evaluation and looks instead at the state specified by the next symbol. If it forms an accepting state, then the lookahead is set to the guard symbol. (If the guard symbol would cause a reduce, we additionally check that the rule will reduce successfully)

--HG--
branch : Parse
extra : convert_revision : svn%3A3666fe7e-a552-0410-bc3c-bd4e4c681ee0/trunk/Parse%401082