Changed the definition of ebnf-item so that the '|' operator applies to more items (ie, (a b | c d) now is equivalent to ((a b) | (c d)) instead of (a (b | c) d))
This screws up the way that the language tree is generated, so this test is failing for now - however, the current way that parenthesized expressions are dealt with, as well as alternates, doesn't work too well for generating the final grammar. The format of the tree really needs to be changed to fix this problem, I think.

--HG--
branch : Parse
extra : convert_revision : svn%3A3666fe7e-a552-0410-bc3c-bd4e4c681ee0/trunk/Parse%401185