Changed the way that weak blocks are declared. Instead of weak-lexer, we now use 'weak lexer' and 'weak keywords', which allows us to fix the warnings with the language definition (where all of the keywords are weak)
Fixed the way the compiler works so that the lexer blocks actually gets compiled. We were previously returning null for the block types that a particular item was not; I've added a 'general' lexer block definition that makes it possible to retrieve a general block and use a switch on the type.

--HG--
branch : Parse
extra : convert_revision : svn%3A3666fe7e-a552-0410-bc3c-bd4e4c681ee0/trunk/Parse%401191