Was dealing with the '%' (end of guard) symbol incorrectly: we were reducing it regardless of the results of can_reduce.
can_reduce was returning the wrong results if the action it was given was not a reduce action. We had a lot of simple guards with an immediate accept action before, so these two bugs cancelled each other out (guards always had a final 'accept' state).

This was the cause of the issue with the ANSI C language where 'int** x;' wasn't accepted as a declarator: we would try to reduce using the '%' symbol in a state where it would not be eventually accepted.

--HG--
branch : Examples