Changed guard behaviour on can_reduce. Previously if can_reduce tried to evaluate a symbol that had a guard on it, it would always return false: this behaviour was mainly because it wasn't clear what the correct behaviour would be.
I've changed this to go on the assumption that the guard is not matched. This makes the parser work when a symbol has both a guard and a weak reduction (previously the guard would always cause the weak reduction check to fail regardless of whether or not the symbol can be reduced or not). The idea is that can_reduce is called with the actual symbol that gets reduced: guard substitutions are assumed to already be applied, so guards do not need to be evaluated further.

This may be non-obvious if can_reduce is called by a user program (perhaps to facilitate error recovery). The solution is probably best dealt with when some more formal error recovery system is implemented.

--HG--
branch : Stable-ish